fix: add t.Parallel() to subtests, move testTransport to helpers, fix Content-Type
- Add t.Parallel() to all 8 table-driven subtests to prevent race conditions
- Move testTransport from describe_instances_test.go to helpers_test.go
  since it's used across 5 test files
- Only set Content-Type header on requests with a body (not GET requests)
