+++ connectors/aws/create_presigned_url_test.go
@@ -96,6 +96,7 @@ func TestCreatePresignedURL_MissingParams(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
 			_, err := action.Execute(t.Context(), connectors.ActionRequest{
 				ActionType:  "aws.create_presigned_url",
 				Parameters:  json.RawMessage(tt.params),
