+++ connectors/aws/get_metrics_test.go
@@ -90,6 +90,7 @@ func TestGetMetrics_MissingParams(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
 			_, err := action.Execute(t.Context(), connectors.ActionRequest{
 				ActionType:  "aws.get_metrics",
 				Parameters:  json.RawMessage(tt.params),
