+++ connectors/aws/list_s3_objects_test.go
@@ -84,6 +84,7 @@ func TestListS3Objects_MissingParams(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
 			_, err := action.Execute(t.Context(), connectors.ActionRequest{
 				ActionType:  "aws.list_s3_objects",
 				Parameters:  json.RawMessage(tt.params),
